Refactor code structure (#962)
move codes and clean spark-2.3 package
