+++ pom.xml
@@ -674,8 +674,8 @@
                 </goals>
                 <configuration>
                   <sources>
-                    <source>src/main/spark2.3.2</source>
-                    <source>src/main/spark2.3</source>
+                    <source>src/main/spark2.3.2/java</source>
+                    <source>src/main/spark2.3.2/scala</source>
                   </sources>
                 </configuration>
               </execution>
@@ -687,7 +687,7 @@
                 </goals>
                 <configuration>
                   <sources>
-                    <source>src/test/spark2.3</source>
+                    <source>src/test/spark2.3/scala</source>
                   </sources>
                 </configuration>
               </execution>
