+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/io/VectorizedCacheReader.scala
