+++ src/main/scala/org/apache/spark/sql/hive/OapSessionStateBuilder.scala
@@ -22,7 +22,7 @@ import org.apache.spark.sql.catalyst.parser.{ParserInterface, SqlBaseParser}
 import org.apache.spark.sql.execution.{SparkPlanner, SparkSqlParser}
 import org.apache.spark.sql.execution.datasources.{DataSourceStrategy, FileSourceStrategy}
 import org.apache.spark.sql.execution.datasources.oap.OapStrategies
-import org.apache.spark.sql.internal.{SQLConf, SessionState, VariableSubstitution}
+import org.apache.spark.sql.internal.{SessionState, SQLConf, VariableSubstitution}
 
 class OapSessionStateBuilder(session: SparkSession, parentState: Option[SessionState] = None)
   extends HiveSessionStateBuilder(session, parentState) {
