+++ src/main/scala/org/apache/spark/sql/oap/adapter/AggregateFunctionAdapter.scala
