--- src/test/spark2.3/scala/org/apache/spark/sql/execution/datasources/oap/adapter/PropertiesAdapter.scala
@@ -1,29 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.execution.datasources.oap.adapter
-
-import org.scalacheck.{Prop, Properties}
-
-object PropertiesAdapter {
-  /**
-    * The Properties has changed in the spark2.3 version.
-    * Ignore it in the spark2.1, spark2.2 version
-    */
-  def getProp(properties: Properties): Prop = Prop.all(properties.properties.map(_._2): _*)
-
-}
